float32

+++ imgseg/annotationUtils.py
@@ -560,6 +560,6 @@ class WeightedBoarderGenerator(MaskGenerator):
 
         # Note: saved as float 16 - to plot has to be converted to float32
         # To be saved rescaled as 8 bit
-        mask_dict['weighted_boarder'] = msk.astype('float16')
+        mask_dict['weighted_boarder'] = msk.astype('float32')
 
         return mask_dict
